1.basic-agent/pyproject.toml
[project]
name = "1-basic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents[litellm]>=0.0.16",
    "python-dotenv>=1.1.0",
]
